ENH: [WIP] RieszFrequencyFunction returns complex.
std::log2(x) substituted by log(x)/vnl_math::ln2 in VowWavelet
std::tie removed.
initializer list in itkInd2SubTest changed to push_back
std:pow(x, unsigned int) to std::pow(static_cast<double>(x),
static_cast<int>(unsigned)). Solve error in gcc4.2
Remove unnecessary typename in tests.

+++ Modules/Filtering/IsotropicWavelets/include/itkHeldIsotropicWavelet.hxx
@@ -55,8 +55,8 @@ typename HeldIsotropicWavelet<TFunctionValue, VImageDimension, TInput>::Function
 HeldIsotropicWavelet<TFunctionValue, VImageDimension, TInput>::EvaluateForwardLowPassFilter(
   const FunctionValueType & freq_norm_in_hz) const
 {
-  FunctionValueType value =
-    std::pow(freq_norm_in_hz, this->m_HighPassSubBands) * std::pow(2.0, 2 * this->m_HighPassSubBands - 1);
+  FunctionValueType value = std::pow(freq_norm_in_hz, static_cast<int>(this->m_HighPassSubBands)) *
+                            std::pow(2.0, static_cast<int>(2 * this->m_HighPassSubBands - 1));
   if (value > 0.25)
     return this->EvaluateMagnitude(value);
   return 1;
@@ -67,8 +67,8 @@ typename HeldIsotropicWavelet<TFunctionValue, VImageDimension, TInput>::Function
 HeldIsotropicWavelet<TFunctionValue, VImageDimension, TInput>::EvaluateForwardHighPassFilter(
   const FunctionValueType & freq_norm_in_hz) const
 {
-  FunctionValueType value =
-    std::pow(freq_norm_in_hz, this->m_HighPassSubBands) * std::pow(2.0, this->m_HighPassSubBands - 1);
+  FunctionValueType value = std::pow(freq_norm_in_hz, static_cast<int>(this->m_HighPassSubBands)) *
+                            std::pow(2.0, static_cast<int>(this->m_HighPassSubBands - 1));
   if (value < 0.25)
     return this->EvaluateMagnitude(value);
   return 1;
@@ -86,8 +86,8 @@ HeldIsotropicWavelet<TFunctionValue, VImageDimension, TInput>::EvaluateForwardSu
     return this->EvaluateForwardLowPassFilter(freq_norm_in_hz);
   if (j > this->m_HighPassSubBands || j < 0)
     throw itk::ExceptionObject(__FILE__, __LINE__, "Invalid SubBand", ITK_LOCATION);
-  FunctionValueType value =
-    std::pow(freq_norm_in_hz, this->m_HighPassSubBands) * std::pow(2.0, 2 * this->m_HighPassSubBands - 1 - j);
+  FunctionValueType value = std::pow(freq_norm_in_hz, static_cast<int>(this->m_HighPassSubBands)) *
+                            std::pow(2.0, static_cast<int>(2 * this->m_HighPassSubBands - 1 - j));
   return this->EvaluateMagnitude(value);
 }
